+++ format_test.go
@@ -1,5 +1,5 @@
 // logger: simple and opinionated log/Slog.Logger instance creator
-// Copyright 2024-2025 by authors and contributors (see AUTHORS file)
+// Copyright 2024-2026 by authors and contributors (see AUTHORS file)
 
 package logger_test
 
@@ -79,6 +79,34 @@ func TestFormat_invalid(t *testing.T) {
 		t.Run(fmt.Sprintf("format=%q", tc), func(t *testing.T) {
 			expected := ""
 			actual, err := logger.Format(tc)
+
+			require.Equal(t, expected, actual)
+
+			if assert.Error(t, err) {
+				require.Equal(t, err, fmt.Errorf("invalid log format: %s", tc))
+			}
+		})
+	}
+}
+
+func TestNormalizeFormat_valid(t *testing.T) {
+	for _, tc := range formats_valid {
+		t.Run(fmt.Sprintf("format=%q", tc), func(t *testing.T) {
+			expected := strings.ToUpper(tc)
+			actual, err := logger.NormalizeFormat(tc)
+
+			require.Equal(t, expected, actual)
+
+			require.NoError(t, err)
+		})
+	}
+}
+
+func TestNormalizeFormat_invalid(t *testing.T) {
+	for _, tc := range formats_invalid {
+		t.Run(fmt.Sprintf("format=%q", tc), func(t *testing.T) {
+			expected := ""
+			actual, err := logger.NormalizeFormat(tc)
 
 			require.Equal(t, expected, actual)