+++ level_test.go
@@ -1,11 +1,12 @@
 // logger: simple and opinionated log/Slog.Logger instance creator
-// Copyright 2024-2025 by authors and contributors (see AUTHORS file)
+// Copyright 2024-2026 by authors and contributors (see AUTHORS file)
 
 package logger_test
 
 import (
 	"fmt"
 	"log/slog"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -109,6 +110,39 @@ func TestLevel_invalid(t *testing.T) {
 			// The expected level is INFO, which happily coincides with the
 			// zero value; however, this is actually enforced by the
 			// implementation.
+			require.Equal(t, expected, actual)
+
+			if assert.Error(t, err) {
+				require.Equal(t, err, fmt.Errorf("invalid log level: %s", tc))
+			}
+		})
+	}
+}
+
+func TestNormalizeLevel_valid(t *testing.T) {
+	for _, tc := range levels_valid {
+		t.Run(fmt.Sprintf("level=%q", tc.name), func(t *testing.T) {
+			expected := strings.ToUpper(tc.name)
+			actual, err := logger.NormalizeLevel(tc.name)
+
+			require.Equal(t, expected, actual)
+
+			require.NoError(t, err)
+		})
+	}
+}
+
+func TestNormalizeLevel_invalid(t *testing.T) {
+	for _, tc := range levels_invalid {
+		t.Run(fmt.Sprintf("level=%q", tc), func(t *testing.T) {
+			expected := ""
+			actual, err := logger.NormalizeLevel(tc)
+
+			// Except in some cases, hopefully properly documented, it is
+			// expected the caller to ignore the return value when the
+			// returned error is not nil. This test, however, makes sure
+			// that the internal implementation still complies with the
+			// library's interface in this case.
 			require.Equal(t, expected, actual)
 
 			if assert.Error(t, err) {